# Conflicts:
#	Script/Game/Event/S_FriendData.cs
#	Script/Game/Event/S_FriendRequest.cs

+++ Script/Game/Event/C_Chat.cs
@@ -6,7 +6,7 @@ namespace SimulFactory.Game.Event
     {
         public static void ChatC(Dictionary<byte,object> param)
         {
-            Define.CHAT_TYPE chatType = (Define.CHAT_TYPE)(int)param[0];
+            Define.CHAT_TYPE chatType = (Define.CHAT_TYPE)(long)param[0];
             string sendUser = (string)param[1];
             string chatText = (string)param[2];
             string targetName = (string)param[3];   // 타입이 whisper가 아닌 경우 "" 로 기입되어 옴
